+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/common/AzureArtifact.java
@@ -6,13 +6,12 @@
 package com.microsoft.azure.toolkit.intellij.common;
 
 import com.intellij.icons.AllIcons;
-import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.application.ReadAction;
 import com.intellij.openapi.externalSystem.model.project.ExternalProjectPojo;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.module.ModuleUtil;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.roots.ProjectFileIndex;
-import com.intellij.openapi.util.Computable;
 import com.intellij.openapi.vfs.LocalFileSystem;
 import com.intellij.openapi.vfs.VfsUtil;
 import com.intellij.openapi.vfs.VirtualFile;
@@ -129,7 +128,7 @@ public class AzureArtifact {
         if (this.getReferencedObject() == null) {
             return null;
         }
-        return ApplicationManager.getApplication().runReadAction((Computable<Module>) () -> switch (type) {
+        return ReadAction.nonBlocking(() -> switch (type) {
             case Gradle -> {
                 final Path path = Paths.get(((ExternalProjectPojo) referencedObject).getPath());
                 yield Optional.ofNullable(VfsUtil.findFile(path, true))
@@ -141,7 +140,8 @@ public class AzureArtifact {
                 .map(p -> VfsUtil.findFile(p, true))
                 .map(f -> ProjectFileIndex.getInstance(project).getModuleForFile(f)).orElse(null);
             case File -> getNearestParentModuleForFile((VirtualFile) this.getReferencedObject());
-        });
+        }).expireWhen(project::isDisposed)
+          .executeSynchronously();
     }
 
     @Nullable